Fix rename prompt with prepopulated value
The second argument for `read-string` is `INITIAL-INPUT` which is used
to pre-populate the minibuffer, the fourth `DEFAULT-VALUE` is what it
returns in case the prompt has an empty string.

While the docs for `read-string` say `INITIAL-INPUT` has been superseded
by `DEFAULT-VALUE`, this is misleading as `DEFAULT-VALUE` by itself does
not pre-populate the minibuffer.

Example code:

```elisp
;; No pre-completed value, returns "" when empty.
(read-string "PROMPT ")

;; Pre-completed with "A", returns "" when empty.
(read-string "PROMPT " "A")

;; Pre-completed with "A", returns "B" when empty.
(read-string "PROMPT " "A" nil "B")

;; No pre-completed value, returns "B" when empty.
(read-string "PROMPT " nil nil "B")
```